Simplified the core fitnessService classes to only show important GENN override
