+++ example/full_visual_example/lib/fitness_services/logical_xor/logical_xor_output.dart
@@ -1,0 +1,6 @@
+/// The possible outputs for the Logical XOR Neural etwork
+enum LogicalXorOutput {
+  yes,
+  no,
+  unsure,
+}
