+++ example/full_visual_example/lib/fitness_services/number_classifier/number_classifier_fitness_service.dart
@@ -1,14 +1,21 @@
 import 'dart:math';
 
-import 'package:flutter/material.dart';
 import 'package:full_visual_example/fitness_services/number_classifier/natural_number.dart';
-import 'package:full_visual_example/visualization_helpers/genn_visualization_example/genn_visualization_example_fitness_service.dart';
+import 'package:full_visual_example/fitness_services/number_classifier/number_classifier_fitness_service_helpers.dart';
+import 'package:full_visual_example/visualization_helpers/genn_visualization_example/visualization_example_genn_fitness_service.dart';
 import 'package:genetically_evolving_neural_network/genetically_evolving_neural_network.dart';
 
 /// This class will be used to score a Number Classifier in tandem with a
 /// Neural Network.
 class NumberClassifierFitnessService
-    extends GENNVisualizationExampleFitnessService<PixelImage, NaturalNumber> {
+    // NOTE: We are extending [GENNFitnessServiceVisualizationExample] instead
+    //       of [GENNFitnessService] to include additional functions to assist
+    //       with visualizing this example project.
+    extends VisualizationExampleGENNFitnessService<PixelImage, NaturalNumber>
+    // NOTE: This provides implementations of the above
+    //       [GENNFitnessServiceVisualizationExample] class.
+    with
+        NumberClassifierFitnessServiceHelpers {
 // ================== START OF GENN EXAMPLE RELATED CONTENT =============================
 // ================== GENNFitnessService Overrides ======================================
 
@@ -45,118 +52,4 @@ class NumberClassifierFitnessService
     return pow(4, points).toDouble();
   }
 // ================== END OF GENN EXAMPLE RELATED CONTENT ===============================
-
-  @override
-  String convertToReadableString(NaturalNumber value) {
-    return NaturalNumber.values.indexOf(value).toString();
-  }
-
-  @override
-  double? get highestPossibleScore =>
-      pow(4, NaturalNumber.values.length).toDouble();
-
-  @override
-  List<PixelImage> get inputList => NaturalNumber.values
-      .map(
-        (naturalNumber) => naturalNumber.asPixelImage(),
-      )
-      .toList();
-
-  /// Returns a list of the integers 0 to 9 represented as pixels on a grid.
-  @override
-  List<Widget> get readableInputList {
-    const tileSideLength = 3.4; // Hardcoded to look nice on the page
-
-    // Creates a Grid View representation of the Natural Number to input
-    return inputList.map((logicalInput) {
-      return Padding(
-        padding: const EdgeInsets.all(1.5),
-        child: Container(
-          color: Colors.blue.withOpacity(0.2),
-          child: SizedBox(
-            width: tileSideLength * 3,
-            height: tileSideLength * 5,
-            child: Padding(
-              padding: const EdgeInsets.all(0.5),
-              child: GridView.count(
-                crossAxisSpacing: 1.0,
-                mainAxisSpacing: 1.0,
-                crossAxisCount: 3,
-                children: List.generate(
-                  logicalInput.pixels.length,
-                  (index) => Container(
-                    color: Colors.black.withOpacity(
-                      logicalInput.pixels[index],
-                    ),
-                    height: tileSideLength,
-                    width: tileSideLength,
-                  ),
-                ),
-              ),
-            ),
-          ),
-        ),
-      );
-    }).toList();
-  }
-
-  @override
-  List<Widget> get readableTargetList => targetOutputsList
-      .map(
-        (targetOutput) => Text(
-          convertToReadableString(targetOutput),
-        ),
-      )
-      .toList();
-
-  @override
-  List<NaturalNumber> get targetOutputsList => NaturalNumber.values;
-
-  @override
-  int get numInitialInputs => NaturalNumberExtension.numPixels;
-
-  @override
-  int get numOutputs => NaturalNumber.values.length;
-
-  /// Returns the index of the guess with the highest confidence.
-  ///
-  @override
-  NaturalNumber convertGuessToOutputType({required List<double> guess}) {
-    // Declare a list to hold the indices of the highest confidence guesses
-    final List<int> listOfHighestConfidenceIndices = [];
-
-    // Declare our initial confidence of 0, because we're not confident in
-    // anything yet.
-    double confidence = 0;
-
-    // Loop through each guess
-    for (int i = 0; i < guess.length; i++) {
-      if (guess[i] > confidence) {
-        // If the current guess has a higher confidence, set it accordingly
-        confidence = guess[i];
-        // Clear any previous indicies of highest confidence
-        listOfHighestConfidenceIndices.clear();
-        listOfHighestConfidenceIndices.add(i);
-      } else if (guess[i] == confidence) {
-        // If the current guess has the same confidence as the current highest
-        listOfHighestConfidenceIndices.add(i);
-      }
-    }
-
-    // Choose an index if there were any ties.
-    int highestConfidenceIndex = accountForGuessTies(
-      listOfHighestConfidenceIndices: listOfHighestConfidenceIndices,
-    );
-
-    return NaturalNumberExtension.parse(
-      number: highestConfidenceIndex,
-    );
-  }
-
-  @override
-  List<double> convertInputToNeuralNetworkInput({
-    required PixelImage input,
-  }) {
-    return input.pixels;
-  }
 }
